# web/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Mamao.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:33191"
    }
  ],
  "paths": {
...
    "/api/v1/employees": {
      "get": {
        "tags": [
          "Employees"
        ],
        "operationId": "listEmployees",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "name": "departmentId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfEmployeeListItem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Employees"
        ],
        "operationId": "createEmployee",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/employees/{id}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "operationId": "getEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Employees"
        ],
        "operationId": "updateEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/employees/{id}/terminate": {
      "post": {
        "tags": [
          "Employees"
        ],
        "operationId": "terminateEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminateEmployeeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/employees/import/preview": {
      "post": {
        "tags": [
          "Employees"
        ],
        "operationId": "previewEmployeeImport",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "arquivo"
                ],
                "type": "object",
                "properties": {
                  "arquivo": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeImportPreview"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/employees/import": {
      "post": {
        "tags": [
          "Employees"
        ],
        "operationId": "importEmployees",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "arquivo"
                ],
                "type": "object",
                "properties": {
                  "arquivo": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeImportResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEmployeeRequest": {
        "required": [
          "fullName",
          "positionId",
          "hiredOn",
          "code",
          "departmentId",
          "managerId"
        ],
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "positionId": {
            "$ref": "#/components/schemas/PositionId"
          },
          "hiredOn": {
            "type": "string",
            "format": "date"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "departmentId": {
            "type": "string",
            "oneOf": [
              {
                "$ref": "#/components/schemas/DepartmentId"
              }
            ],
            "nullable": true
          },
          "managerId": {
            "type": "string",
            "oneOf": [
              {
                "$ref": "#/components/schemas/EmployeeId"
              }
            ],
            "nullable": true
          }
        }
      },
...
      "DepartmentId": {
        "type": "string",
        "format": "uuid"
      },
...
      "EmployeeId": {
        "type": "string",
        "format": "uuid"
      },
      "EmployeeImportColumn": {
        "required": [
          "field",
          "label",
          "header",
          "required",
          "acceptedNames"
        ],
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "header": {
            "type": "string",
            "nullable": true
          },
          "required": {
            "type": "boolean"
          },
          "acceptedNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "EmployeeImportPreview": {
        "required": [
          "columns",
          "rows",
          "summary",
          "warnings"
        ],
        "type": "object",
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmployeeImportColumn"
            }
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmployeeImportRow"
            }
          },
          "summary": {
            "$ref": "#/components/schemas/EmployeeImportSummary"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "EmployeeImportResult": {
        "required": [
          "imported",
          "skipped"
        ],
        "type": "object",
        "properties": {
          "imported": {
            "format": "int32",
            "type": "integer"
          },
          "skipped": {
            "format": "int32",
            "type": "integer"
          }
        }
      },
      "EmployeeImportRow": {
        "required": [
          "lineNumber",
          "code",
          "fullName",
          "positionName",
          "hiredOn",
          "status",
          "errors"
        ],
        "type": "object",
        "properties": {
          "lineNumber": {
            "format": "int32",
            "type": "integer"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string"
          },
          "positionName": {
            "type": "string"
          },
          "hiredOn": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/EmployeeImportRowStatus"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "EmployeeImportRowStatus": {
        "enum": [
          "Pronta",
          "Invalida",
          "DuplicadaNoArquivo",
          "DuplicadaNoCadastro"
        ]
      },
      "EmployeeImportSummary": {
        "required": [
          "total",
          "ready",
          "invalid",
          "duplicates"
        ],
        "type": "object",
        "properties": {
          "total": {
            "format": "int32",
            "type": "integer"
          },
          "ready": {
            "format": "int32",
            "type": "integer"
          },
          "invalid": {
            "format": "int32",
            "type": "integer"
          },
          "duplicates": {
            "format": "int32",
            "type": "integer"
          }
        }
      },
      "EmployeeListItem": {
        "required": [
          "id",
          "code",
          "fullName",
          "positionName",
          "departmentName",
          "hiredOn",
          "isActive"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/EmployeeId"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string"
          },
          "positionName": {
            "type": "string"
          },
          "departmentName": {
            "type": "string",
            "nullable": true
          },
          "hiredOn": {
            "type": "string",
            "format": "date"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "EmployeeResponse": {
        "required": [
          "id",
          "code",
          "fullName",
          "positionId",
          "positionName",
          "departmentId",
          "departmentName",
          "managerId",
          "managerName",
          "hiredOn",
          "terminatedOn",
          "isActive",
          "hasLogin"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/EmployeeId"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string"
          },
          "positionId": {
            "$ref": "#/components/schemas/PositionId"
          },
          "positionName": {
            "type": "string"
          },
          "departmentId": {
            "type": "string",
            "oneOf": [
              {
                "$ref": "#/components/schemas/DepartmentId"
              }
            ],
            "nullable": true
          },
          "departmentName": {
            "type": "string",
            "nullable": true
          },
          "managerId": {
            "type": "string",
            "oneOf": [
              {
                "$ref": "#/components/schemas/EmployeeId"
              }
            ],
            "nullable": true
          },
          "managerName": {
            "type": "string",
            "nullable": true
          },
          "hiredOn": {
            "type": "string",
            "format": "date"
          },
          "terminatedOn": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "hasLogin": {
            "type": "boolean"
          }
        }
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "PagedResultOfEmployeeListItem": {
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmployeeListItem"
            }
          },
          "total": {
            "format": "int32",
            "type": "integer"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "pageSize": {
            "format": "int32",
            "type": "integer"
          }
        }
      },
      "PositionId": {
        "type": "string",
        "format": "uuid"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TerminateEmployeeRequest": {
        "required": [
          "terminatedOn"
        ],
        "type": "object",
        "properties": {
          "terminatedOn": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "UpdateEmployeeRequest": {
        "required": [
          "fullName",
          "positionId",
          "departmentId",
          "managerId"
        ],
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "positionId": {
            "$ref": "#/components/schemas/PositionId"
          },
          "departmentId": {
            "type": "string",
            "oneOf": [
              {
                "$ref": "#/components/schemas/DepartmentId"
              }
            ],
            "nullable": true
          },
          "managerId": {
            "type": "string",
            "oneOf": [
              {
                "$ref": "#/components/schemas/EmployeeId"
              }
            ],
            "nullable": true
          }
        }
      },
...
    }
  },
...
}